backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pymongo==4.7.2
python-dotenv==1.0.1
scikit-learn==1.5.0
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
joblib==1.4.2
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
requests==2.31.0
